requirements.txt
PyOpenGL>=3.1.0
opencv-python>=4.1.1.26
numpy>=1.17.3